feat: move account details and privacy to a dedicated Account settings page
Extracts the Account Details and Privacy & Subscription sections out of the settings Profile card into a new AccountSettings component, routed at /settings/account and listed below General in the settings navigation for both desktop and mobile. Privacy & Subscription rows are merged into the Account Details card.
