+++ projects/client/i18n/messages/ar-SA.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "إيجابي",
   "header_settings_clear_cover": "مسح صورة الغلاف",
   "header_settings_notifications": "الإخطارات",
-  "header_settings_privacy": "الخصوصية والاشتراك",
   "header_settings_tracking": "تتبع",
   "header_settings_translations": "الترجمات",
   "header_skipped_items": "العناصر التي تم تخطيها ({count})",
