+++ projects/client/i18n/messages/bg-BG.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Положително",
   "header_settings_clear_cover": "Изчисти корицата",
   "header_settings_notifications": "Известия",
-  "header_settings_privacy": "Поверителност & Абонамент",
   "header_settings_tracking": "Проследяване",
   "header_settings_translations": "Преводи",
   "header_skipped_items": "{count} пропуснати елемента",
