+++ projects/client/i18n/messages/ca-ES.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Positiu",
   "header_settings_clear_cover": "Esborra la imatge de portada",
   "header_settings_notifications": "Notificacions",
-  "header_settings_privacy": "Privadesa i subscripció",
   "header_settings_tracking": "Seguiment",
   "header_settings_translations": "Traduccions",
   "header_skipped_items": "{count} elements omesos",
