+++ projects/client/i18n/messages/da-DK.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Positiv",
   "header_settings_clear_cover": "Ryd Coverbillede",
   "header_settings_notifications": "Notifikationer",
-  "header_settings_privacy": "Privatliv & Abonnement",
   "header_settings_tracking": "Sporing",
   "header_settings_translations": "Oversættelser",
   "header_skipped_items": "{count} Oversprungne Elementer",
