+++ projects/client/i18n/messages/de-DE.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Positiv",
   "header_settings_clear_cover": "Coverbild löschen",
   "header_settings_notifications": "Benachrichtigungen",
-  "header_settings_privacy": "Datenschutz & Abonnement",
   "header_settings_tracking": "Tracking",
   "header_settings_translations": "Übersetzungen",
   "header_skipped_items": "{count} übersprungene Elemente",
