+++ projects/client/i18n/messages/el-GR.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Θετικό",
   "header_settings_clear_cover": "Εκκαθάριση εικόνας εξωφύλλου",
   "header_settings_notifications": "Ειδοποιήσεις",
-  "header_settings_privacy": "Απόρρητο & Συνδρομή",
   "header_settings_tracking": "Παρακολούθηση",
   "header_settings_translations": "Μεταφράσεις",
   "header_skipped_items": "{count} Παραλειφθέντα στοιχεία",
