+++ projects/client/i18n/messages/en-AU.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Positive",
   "header_settings_clear_cover": "Clear Cover Image",
   "header_settings_notifications": "Notifications",
-  "header_settings_privacy": "Privacy & Subscription",
   "header_settings_tracking": "Tracking",
   "header_settings_translations": "Translations",
   "header_skipped_items": "{count} Skipped Items",
