+++ projects/client/i18n/messages/es-MX.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Positivo",
   "header_settings_clear_cover": "Borrar imagen de portada",
   "header_settings_notifications": "Notificaciones",
-  "header_settings_privacy": "Privacidad y Suscripción",
   "header_settings_tracking": "Seguimiento",
   "header_settings_translations": "Traducciones",
   "header_skipped_items": "{count} elementos omitidos",
