+++ projects/client/i18n/messages/fa-IR.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "مثبت",
   "header_settings_clear_cover": "پاک کردن تصویر جلد",
   "header_settings_notifications": "اعلان‌ها",
-  "header_settings_privacy": "حریم خصوصی و اشتراک",
   "header_settings_tracking": "ردیابی",
   "header_settings_translations": "ترجمه‌ها",
   "header_skipped_items": "{count} مورد نادیده گرفته شده",
