+++ projects/client/i18n/messages/fr-FR.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Positif",
   "header_settings_clear_cover": "Effacer l'image de couverture",
   "header_settings_notifications": "Notifications",
-  "header_settings_privacy": "Confidentialité et abonnement",
   "header_settings_tracking": "Suivi",
   "header_settings_translations": "Traductions",
   "header_skipped_items": "{count} éléments ignorés",
