+++ projects/client/i18n/messages/hu-HU.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Pozitív",
   "header_settings_clear_cover": "Borítókép törlése",
   "header_settings_notifications": "Értesítések",
-  "header_settings_privacy": "Adatvédelem és Előfizetés",
   "header_settings_tracking": "Nyomon követés",
   "header_settings_translations": "Fordítások",
   "header_skipped_items": "{count} kihagyott elem",
