+++ projects/client/i18n/messages/id-ID.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Positif",
   "header_settings_clear_cover": "Hapus Gambar Sampul",
   "header_settings_notifications": "Notifikasi",
-  "header_settings_privacy": "Privasi & Langganan",
   "header_settings_tracking": "Pelacakan",
   "header_settings_translations": "Terjemahan",
   "header_skipped_items": "{count} Item Dilewati",
