+++ projects/client/i18n/messages/it-IT.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Positivo",
   "header_settings_clear_cover": "Cancella immagine di copertina",
   "header_settings_notifications": "Notifiche",
-  "header_settings_privacy": "Privacy e Abbonamento",
   "header_settings_tracking": "Tracciamento",
   "header_settings_translations": "Traduzioni",
   "header_skipped_items": "{count} elementi saltati",
