+++ projects/client/i18n/messages/ja-JP.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "ポジティブ",
   "header_settings_clear_cover": "カバー画像をクリア",
   "header_settings_notifications": "通知",
-  "header_settings_privacy": "プライバシーとサブスクリプション",
   "header_settings_tracking": "トラッキング",
   "header_settings_translations": "翻訳",
   "header_skipped_items": "{count}個のスキップ済みアイテム",
