+++ projects/client/i18n/messages/nb-NO.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Positiv",
   "header_settings_clear_cover": "Fjern forsidebilde",
   "header_settings_notifications": "Varsler",
-  "header_settings_privacy": "Personvern og abonnement",
   "header_settings_tracking": "Sporing",
   "header_settings_translations": "Oversettelser",
   "header_skipped_items": "{count} elementer hoppet over",
