+++ projects/client/i18n/messages/nl-NL.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Positief",
   "header_settings_clear_cover": "Omslagafbeelding wissen",
   "header_settings_notifications": "Meldingen",
-  "header_settings_privacy": "Privacy & Abonnement",
   "header_settings_tracking": "Bijhouden",
   "header_settings_translations": "Vertalingen",
   "header_skipped_items": "{count} overgeslagen items",
