+++ projects/client/i18n/messages/pl-PL.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Pozytywny",
   "header_settings_clear_cover": "Wyczyść obraz tła",
   "header_settings_notifications": "Powiadomienia",
-  "header_settings_privacy": "Prywatność i subskrypcja",
   "header_settings_tracking": "Śledzenie",
   "header_settings_translations": "Tłumaczenia",
   "header_skipped_items": "{count} pominiętych elementów",
