+++ projects/client/i18n/messages/pt-BR.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Positivo",
   "header_settings_clear_cover": "Limpar Imagem de Capa",
   "header_settings_notifications": "Notificações",
-  "header_settings_privacy": "Privacidade e Assinatura",
   "header_settings_tracking": "Rastreamento",
   "header_settings_translations": "Traduções",
   "header_skipped_items": "{count} Itens Ignorados",
