+++ projects/client/i18n/messages/pt-PT.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Positivo",
   "header_settings_clear_cover": "Remover imagem de capa",
   "header_settings_notifications": "Notificações",
-  "header_settings_privacy": "Privacidade e subscrição",
   "header_settings_tracking": "Registo",
   "header_settings_translations": "Traduções",
   "header_skipped_items": "{count} itens ignorados",
