+++ projects/client/i18n/messages/ro-RO.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Pozitiv",
   "header_settings_clear_cover": "Șterge Imaginea de Copertă",
   "header_settings_notifications": "Notificări",
-  "header_settings_privacy": "Confidențialitate și Abonament",
   "header_settings_tracking": "Urmărire",
   "header_settings_translations": "Traduceri",
   "header_skipped_items": "{count} Elemente Omise",
