+++ projects/client/i18n/messages/ru-RU.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Положительный",
   "header_settings_clear_cover": "Очистить обложку",
   "header_settings_notifications": "Уведомления",
-  "header_settings_privacy": "Конфиденциальность и подписка",
   "header_settings_tracking": "Отслеживание",
   "header_settings_translations": "Переводы",
   "header_skipped_items": "{count} пропущенных элементов",
