+++ projects/client/i18n/messages/sv-SE.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Positiv",
   "header_settings_clear_cover": "Rensa omslagsbild",
   "header_settings_notifications": "Aviseringar",
-  "header_settings_privacy": "Integritet och prenumeration",
   "header_settings_tracking": "Spårning",
   "header_settings_translations": "Översättningar",
   "header_skipped_items": "{count} överhoppade objekt",
