+++ projects/client/i18n/messages/tr-TR.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Olumlu",
   "header_settings_clear_cover": "Kapak Resmini Temizle",
   "header_settings_notifications": "Bildirimler",
-  "header_settings_privacy": "Gizlilik ve Abonelik",
   "header_settings_tracking": "Takip",
   "header_settings_translations": "Çeviriler",
   "header_skipped_items": "{count} Atlanan Öğe",
