+++ projects/client/i18n/messages/uk-UA.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "Позитивний",
   "header_settings_clear_cover": "Очистити обкладинку",
   "header_settings_notifications": "Сповіщення",
-  "header_settings_privacy": "Приватність і Підписка",
   "header_settings_tracking": "Відстеження",
   "header_settings_translations": "Переклади",
   "header_skipped_items": "{count} пропущених елементів",
