+++ projects/client/i18n/messages/zh-CN.json
@@ -727,7 +727,6 @@
   "header_sentiment_positive": "积极",
   "header_settings_clear_cover": "清除封面图片",
   "header_settings_notifications": "通知",
-  "header_settings_privacy": "隐私与订阅",
   "header_settings_tracking": "追踪",
   "header_settings_translations": "翻译",
   "header_skipped_items": "{count} 个已跳过项目",
