+++ projects/client/i18n/meta/en.json
@@ -3893,10 +3893,6 @@
       "default": "Account Details",
       "description": "Header for the section that allows users to change their account details."
     },
-    "header_settings_privacy": {
-      "default": "Privacy & Subscription",
-      "description": "Header for the settings section covering privacy options and subscription management."
-    },
     "description_account_details": {
       "default": "Change your photo, name and other details.",
       "description": "Description for the account details section."
@@ -4121,6 +4117,10 @@
     "link_text_general_settings": {
       "default": "General",
       "description": "Text for the link to the general settings page."
+    },
+    "link_text_account_settings": {
+      "default": "Account",
+      "description": "Text for the link to the account settings page."
     },
     "link_text_data_settings": {
       "default": "Data",
