+++ projects/client/src/lib/utils/url/UrlBuilder.ts
@@ -314,6 +314,7 @@ export const UrlBuilder = {
   settings: {
     general: () => '/settings',
     generalDetail: () => '/settings/general',
+    account: () => '/settings/account',
     data: () => '/settings/data',
     apps: () => '/settings/apps',
     appsConnected: () => '/settings/apps/connected',
