+++ projects/client/src/routes/settings/account/+page.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+  import AccountSettings from "$lib/sections/settings/AccountSettings.svelte";
+</script>
+
+<AccountSettings />
